Move gamma erp test data to its own file
